requirements.txt
Flask
python-dotenv
gunicorn
requests
flask-cors

